+++ runtime/test/ttnn/gtest/test_tensor_cache_lifetime.cpp
@@ -28,7 +28,7 @@ protected:
 
   tt::runtime::Tensor createTestTensor() {
     std::vector<uint32_t> shape = {2, 2};
-    std::vector<uint32_t> stride = tt::runtime::utils::calculateStride(shape);
+    std::vector<uint64_t> stride = tt::runtime::utils::calculateStride(shape);
     tt::target::DataType dataType = tt::target::DataType::Float32;
     uint32_t itemSize = sizeof(float);
